# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Activepieces Documentation",
    "version": "0.14.3"
  },
...
  "paths": {
    "/v1/app-connections/": {
      "post": {
        "tags": [
          "app-connections"
        ],
        "description": "Upsert an app connection based on the app name",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "title": "Secret Text",
                    "description": "Secret Text",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "SECRET_TEXT"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "SECRET_TEXT"
                            ]
                          },
                          "secret_text": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "type",
                          "secret_text"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  },
                  {
                    "title": "OAuth2",
                    "description": "OAuth2",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "OAUTH2"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "client_id": {
                            "type": "string"
                          },
                          "client_secret": {
                            "type": "string"
                          },
                          "grant_type": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "authorization_code"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "client_credentials"
                                ]
                              }
                            ]
                          },
                          "token_url": {
                            "type": "string"
                          },
                          "props": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "scope": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "authorization_method": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "HEADER"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "BODY"
                                ]
                              }
                            ]
                          },
                          "code_challenge": {
                            "type": "string"
                          },
                          "redirect_url": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "OAUTH2"
                            ]
                          }
                        },
                        "required": [
                          "client_id",
                          "client_secret",
                          "token_url",
                          "scope",
                          "code",
                          "redirect_url",
                          "type"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  },
                  {
                    "title": "Cloud OAuth2",
                    "description": "Cloud OAuth2",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "CLOUD_OAUTH2"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "client_id": {
                            "type": "string"
                          },
                          "authorization_method": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "HEADER"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "BODY"
                                ]
                              }
                            ]
                          },
                          "code": {
                            "type": "string"
                          },
                          "code_challenge": {
                            "type": "string"
                          },
                          "props": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "scope": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "CLOUD_OAUTH2"
                            ]
                          },
                          "token_url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "client_id",
                          "code",
                          "scope",
                          "type"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  },
                  {
                    "title": "Platform OAuth2",
                    "description": "Platform OAuth2",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "PLATFORM_OAUTH2"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "client_id": {
                            "type": "string"
                          },
                          "authorization_method": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "HEADER"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "BODY"
                                ]
                              }
                            ]
                          },
                          "code": {
                            "type": "string"
                          },
                          "code_challenge": {
                            "type": "string"
                          },
                          "props": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "scope": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "PLATFORM_OAUTH2"
                            ]
                          },
                          "token_url": {
                            "type": "string"
                          },
                          "redirect_url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "client_id",
                          "code",
                          "scope",
                          "type",
                          "redirect_url"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  },
                  {
                    "title": "Basic Auth",
                    "description": "Basic Auth",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "BASIC_AUTH"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string"
                          },
                          "password": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "BASIC_AUTH"
                            ]
                          }
                        },
                        "required": [
                          "username",
                          "password",
                          "type"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  },
                  {
                    "title": "Custom Auth",
                    "description": "Custom Auth",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "appName": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "CUSTOM_AUTH"
                        ]
                      },
                      "value": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "CUSTOM_AUTH"
                            ]
                          },
                          "props": {
                            "type": "object",
                            "additionalProperties": {}
                          }
                        },
                        "required": [
                          "type",
                          "props"
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "appName",
                      "projectId",
                      "type",
                      "value"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "tags": [
          "app-connections"
        ],
        "description": "List app connections",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "appName",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "description": "App connection is a connection to an external app.",
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "created": {
                            "type": "string"
                          },
                          "updated": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "OAUTH2"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "PLATFORM_OAUTH2"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "CLOUD_OAUTH2"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "SECRET_TEXT"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "BASIC_AUTH"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "CUSTOM_AUTH"
                                ]
                              }
                            ]
                          },
                          "appName": {
                            "type": "string"
                          },
                          "projectId": {
                            "pattern": "^[0-9a-zA-Z]{21}$",
                            "type": "string"
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "ACTIVE"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "ERROR"
                                ]
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "created",
                          "updated",
                          "name",
                          "type",
                          "appName",
                          "projectId",
                          "status"
                        ]
                      }
                    },
                    "next": {
                      "description": "Cursor to the next page",
                      "type": "string",
                      "nullable": true
                    },
                    "previous": {
                      "description": "Cursor to the previous page",
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/app-connections/{id}": {
      "delete": {
        "tags": [
          "app-connections"
        ],
        "description": "Delete an app connection",
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-zA-Z]{21}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://cloud.activepieces.com/api",
      "description": "Production Server"
    }
  ],
...
}